Switch to monorepo using pnpm. Introduce separate widgets. Create shared ui package.
In the new Zebar, closing widgets has now been fixed. This allows for
fully functional popover windows. Since every widget is a standalone
window, I have now decided to move to a monorepo structure. This will
use pnpm (as it has better support for monorepos). This commit lays the
foundation for this.